+++ packages/api-client/src/v2/blocks/request-block/components/RequestTableRow.vue
@@ -199,6 +199,7 @@ const handleUpdateRow = (
         :modelValue="name"
         placeholder="Key"
         :required="Boolean(data.isRequired)"
+        @navigate="(route) => emit('navigate', route)"
         @selectVariable="(v: string) => handleUpdateRow({ name: v })"
         @update:modelValue="(v) => handleUpdateRow({ name: v })" />
     </DataTableCell>
@@ -225,6 +226,7 @@ const handleUpdateRow = (
         :modelValue="displayValue"
         placeholder="Value"
         :type="typeValue"
+        @navigate="(route) => emit('navigate', route)"
         @update:modelValue="(v) => handleUpdateRow({ value: v })">
         <template #icon>
           <ScalarButton