+++ packages/icons/src/library/LibraryIcon.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { ref, watchEffect, type Component } from 'vue'
+import { shallowRef, watchEffect, type Component } from 'vue'
 
 import { getLibraryIcon } from './icons'
 
@@ -7,7 +7,7 @@ const props = defineProps<{
   src: string
 }>()
 
-const data = ref<Component>()
+const data = shallowRef<Component>()
 
 watchEffect(async () => {
   data.value = await getLibraryIcon(props.src)