+++ packages/workspace-store/src/schemas/extensions/document/x-scalar-icon.ts
@@ -1,4 +1,5 @@
 import { Type } from '@scalar/typebox'
+import { object, optional, string } from '@scalar/validation'
 
 export const XScalarIconSchema = Type.Object({
   'x-scalar-icon': Type.Optional(Type.String()),
@@ -8,3 +9,13 @@ export type XScalarIcon = {
   /** A custom icon representing the collection */
   'x-scalar-icon'?: string
 }
+
+export const XScalarIcon = object(
+  {
+    'x-scalar-icon': optional(string()),
+  },
+  {
+    typeName: 'XScalarIcon',
+    typeComment: 'A custom icon representing the collection',
+  },
+)