+++ integrations/docker/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @scalarapi/docker-api-reference
 
+## 0.5.18
+
+## 0.5.17
+
 ## 0.5.16
 
 ## 0.5.15