+++ packages/api-reference/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @scalar/api-reference
 
+## 1.49.8
+
+### Patch Changes
+
+- [#8628](https://github.com/scalar/scalar/pull/8628): chore: remove unused playgrounds
+
+## 1.49.7
+
+### Patch Changes
+
+- [#8612](https://github.com/scalar/scalar/pull/8612): fix: intersection observer for scroll-spy and tall sections
+
 ## 1.49.6
 
 ### Patch Changes