+++ packages/openapi-parser/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @scalar/openapi-parser
 
+## 0.25.7
+
 ## 0.25.6
 
 ### Patch Changes