+++ packages/validation/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @scalar/validation
 
+## 0.2.0
+
+### Minor Changes
+
+- [#8600](https://github.com/scalar/scalar/pull/8600): feat: first class support for optional and intersection types
+
 ## 0.1.0
 
 ### Minor Changes